requirements.txt
beautifulsoup4==4.8.0
css-html-js-minify==2.5.5
moviepy==1.0.3
Pillow==7.2.0
ffmpeg-python==0.2.0